Refactor classification prompts into package
