+++ backend/src/case_generator/prompts/clasificacion/__init__.py
@@ -1,0 +1,17 @@
+"""Classification-family prompt exports."""
+
+from case_generator.prompts.clasificacion.narrative import (
+    M3_CONTENT_PROMPT_CLASSIFICATION,
+    M4_PROMPT_CLASSIFICATION,
+    M5_PROMPT_CLASSIFICATION,
+)
+from case_generator.prompts.clasificacion.notebook import (
+    M3_NOTEBOOK_ALGO_PROMPT_CLASSIFICATION,
+)
+
+__all__ = [
+    "M3_CONTENT_PROMPT_CLASSIFICATION",
+    "M4_PROMPT_CLASSIFICATION",
+    "M5_PROMPT_CLASSIFICATION",
+    "M3_NOTEBOOK_ALGO_PROMPT_CLASSIFICATION",
+]
